refactor: CSS 경로 변경
